+++ lyr-core/src/main/java/com/lyr/rule/dynamodb/config/ScanDynamodbTableIdleRuleConfig.java
@@ -1,50 +1,27 @@
 package com.lyr.rule.dynamodb.config;
 
-import static com.lyr.util.RuleDefinition.SCAN_DYNAMODB_TABLE_IDLE;
-
-import com.lyr.exception.rule.config.BadRuleConfigException;
-import com.lyr.exception.rule.config.InvalidRuleConfigTypeException;
 import com.lyr.rule.RuleConfig;
 import java.util.Map;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
 import lombok.NonNull;
+import lombok.Value;
 
 @Builder
-@AllArgsConstructor
-@Getter
+@Value
 public class ScanDynamodbTableIdleRuleConfig implements RuleConfig {
     public static final String EXCLUDE_EMPTY_TABLES_CONFIG_KEY = "excludeEmptyTables";
     public static final String MAX_IDLE_PERIOD_IN_DAYS_CONFIG_KEY = "maxIdlePeriodInDays";
 
+    public static final Integer DEFAULT_MAX_IDLE_PERIOD_IN_DAYS = 30;
+    public static final Boolean DEFAULT_EXCLUDE_EMPTY_TABLES = Boolean.FALSE;
+
+    @Builder.Default
     @NonNull
-    private final Integer maxIdlePeriodInDays;
+    Integer maxIdlePeriodInDays = DEFAULT_MAX_IDLE_PERIOD_IN_DAYS;
 
     @NonNull
     @Builder.Default
-    private final Boolean excludeEmptyTables = Boolean.FALSE;
-
-    public static ScanDynamodbTableIdleRuleConfig parse(final Object config) {
-        if (!(config instanceof Map)) {
-            throw new InvalidRuleConfigTypeException(SCAN_DYNAMODB_TABLE_IDLE.getRuleName(), "map");
-        }
-
-        final var configMap = (Map<String, Object>) config;
-
-        final var scanDynamodbTableIdleRuleConfig = ScanDynamodbTableIdleRuleConfig.builder()
-                .maxIdlePeriodInDays(
-                        (Integer) RuleConfig.getMandatoryAttribute(MAX_IDLE_PERIOD_IN_DAYS_CONFIG_KEY, configMap))
-                .excludeEmptyTables((Boolean) configMap.getOrDefault(EXCLUDE_EMPTY_TABLES_CONFIG_KEY, Boolean.FALSE))
-                .build();
-
-        if (scanDynamodbTableIdleRuleConfig.getMaxIdlePeriodInDays() < 1) {
-            throw new BadRuleConfigException("Max idle period attribute of " + SCAN_DYNAMODB_TABLE_IDLE.getRuleName()
-                    + " rule config, must be longer than a day");
-        }
-
-        return scanDynamodbTableIdleRuleConfig;
-    }
+    Boolean excludeEmptyTables = DEFAULT_EXCLUDE_EMPTY_TABLES;
 
     @Override
     public Map<String, String> getConfigAsStringMap() {