+++ lyr-core/src/main/java/com/lyr/rule/lambda/config/ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.java
@@ -2,19 +2,12 @@ package com.lyr.rule.lambda.config;
 
 import com.lyr.rule.RuleConfig;
 import java.util.Map;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Value;
 
 @Builder
-@AllArgsConstructor
-@Getter
+@Value
 public class ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig implements RuleConfig {
-
-    public static ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig parse(final Object ignored) {
-        // No config available for this rule
-        return ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.builder().build();
-    }
 
     @Override
     public Map<String, String> getConfigAsStringMap() {