+++ lyr-core/src/test/java/com/lyr/rule/lambda/config/ScanLambdaFunctionWithUnboundedConcurrencyRuleConfigTest.java
@@ -2,34 +2,13 @@ package com.lyr.rule.lambda.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 
 class ScanLambdaFunctionWithUnboundedConcurrencyRuleConfigTest {
 
     @Test
-    void testThatScanLambdaFunctionWithUnboundedConcurrencyRuleConfigIsParsedSuccessfully() {
-        // Given
-        final Object configList = List.of(15);
-        final Object configMap = Map.of("abc", 15, "def", 30);
-
-        final ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig expectedRuleConfig =
-                ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.builder().build();
-
-        // When
-        final var actualRuleConfigList = ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.parse(configList);
-        final var actualRuleConfigMap = ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.parse(configMap);
-        final var actualRuleConfigNull = ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.parse(null);
-
-        // Then
-        assertThat(actualRuleConfigList).usingRecursiveComparison().isEqualTo(expectedRuleConfig);
-        assertThat(actualRuleConfigMap).usingRecursiveComparison().isEqualTo(expectedRuleConfig);
-        assertThat(actualRuleConfigNull).usingRecursiveComparison().isEqualTo(expectedRuleConfig);
-    }
-
-    @Test
-    void testThatScanLambdaFunctionWithUnboundedConcurrencyRuleConfigIsCopiedCorrectly() {
+    void testThatRuleConfigIsCopiedCorrectly() {
         // Given
         final var expectedConfig =
                 ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.builder().build();
@@ -43,7 +22,7 @@ class ScanLambdaFunctionWithUnboundedConcurrencyRuleConfigTest {
     }
 
     @Test
-    void testThatScanLambdaFunctionWithUnboundedConcurrencyRuleConfigIsConvertedToAMapSuccessfully() {
+    void testThatRuleConfigIsConvertedToAMapSuccessfully() {
         // Given
         final var expectedConfig =
                 ScanLambdaFunctionWithUnboundedConcurrencyRuleConfig.builder().build();